# Speed up fetching range reports, evolution and sparkline reports