--- a/core/Db/Schema/Mysql.php
+++ b/core/Db/Schema/Mysql.php
@@ -298,7 +298,7 @@ class Mysql implements SchemaInterface
                                       ts_archived DATETIME NULL,
                                       value DOUBLE NULL,
                                         PRIMARY KEY(idarchive, name),
-                                        INDEX index_idsite_dates_period(idsite, date1, date2, period, ts_archived),
+                                        INDEX index_idsite_dates_period(idsite, date1, date2, period, name(6)),
                                         INDEX index_period_archived(period, ts_archived)
                                       ) ENGINE=$engine DEFAULT CHARSET=$charset
             ",