--- a/core/Updates/5.0.0-b1.php
+++ b/core/Updates/5.0.0-b1.php
@@ -60,11 +60,12 @@ class Updates_5_0_0_b1 extends PiwikUpdates
 
         $tables = ArchiveTableCreator::getTablesArchivesInstalled('numeric');
         foreach ($tables as $table) {
+            $migrations[] = $this->migration->db->sql(sprintf('DELETE FROM `%s` WHERE ts_archived is null', $table));
+            
             $hasPrefix = strpos($table, 'archive') !== 0;
             if ($hasPrefix) {
                 $table = Common::unprefixTable($table);
             }
-            $migrations[] = $this->migration->db->sql(sprintf('DELETE FROM `%s` WHERE ts_archived is null', $table));
             $migrations[] = $this->migration->db->dropIndex($table, 'index_idsite_dates_period');
             $migrations[] = $this->migration->db->addIndex($table, ['idsite', 'date1', 'date2', 'period', 'name(6)'], 'index_idsite_dates_period');
         }