### Description:

This index change will speed up all kind of reports fetching but especially range periods where the change will be most noticeable. 

When fetching evolution graphs, then the method `getArchiveIds` can be very slow. Especially when looking for many periods.

See eg below where it executed 30 queries and it took 7 seconds:

<img width="874" alt="image" src="https://user-images.githubusercontent.com/273120/208315369-f2343d9b-a7e4-4935-8b82-efcc51951f90.png">

Or this range period requests where it took 217 seconds for 2240 calls (there might be still more tweaks possible to reduce the number of calls which be a separate fix).

<img width="871" alt="image" src="https://user-images.githubusercontent.com/273120/208315423-4c93dafe-4c9c-4e0e-a417-2cfee23d1953.png">

I've been testing this on two of our instances with different queries and profiling and we can reproduce this.

I've then looked at the current numeric index which is:

```sql
`idsite`,`date1`,`date2`,`period`,`ts_archived`
```

The `ts_archived` is not very useful for an index though because most of the queries wouldn't lower the result set. By changing this part of the index to `name(6)`. I was able to make the queries 150-170 times faster on our instances (we ran the same queries 100 times with `SQL_NO_CACHE` and compared different queries with different indexes). That's roughly because our archives include 166 different record names. On some smaller instance it improved the query 40+.

On our account 1 such query takes typically 200ms vs with this index change it takes 1.2ms. 

That means above screenshot goes roughly down from 7 seconds to say 50ms. The 217 seconds goes down to 1.5 seconds.

While not changing the index size in the end as I've looked over the queries containing `ts_archived` and the `ts_archived` index typically doesn't help much there but even more importantly these queries contain also the `name` column meaning the new index should be even more effective there.

This query `getArchiveIds()` is also the TOP contributor on our cloud for load. 

`ts_archived` was added initially 12 years ago in https://github.com/matomo-org/matomo/commit/f593a3e997de85c40cfab123dcd0d13b15009d7d but a better choice would have been to use `name` column but that time there were likely certain MySQL limitations. 

**Why `name(6)`?**
The done flag we use in those archive queries are typically `done`, `done.VisitsSummary` or `done2499999` or `done249999.VisitsSummary`. By including the first 6 characters in the index we should be able to select more precise the best rows. Technically, `name(5)` would do an equally good job as out of 1M records there are only 6 `done.*` records but figured rather give it one more.

**Other changes**

* I've also removed where query parts `ts_archived is not null` as it is always not null. I assume it was added for old MySQL versions as it otherwise wouldn't use the index unless all parts were covered.

We have tested these changes on production on our account and reports are now loading pretty much instant.

### Review

* [ ] [Functional review done](https://developer.matomo.org/guides/pull-request-reviews#functional-review-done)
* [ ] [Potential edge cases thought about](https://developer.matomo.org/guides/pull-request-reviews#potential-edge-cases-thought-about) (behavior of the code with strange input, with strange internal state or possible interactions with other Matomo subsystems)
* [ ] [Usability review done](https://developer.matomo.org/guides/pull-request-reviews#usability-review-done) (is anything maybe unclear or think about anything that would cause people to reach out to support)
* [ ] [Security review done](https://developer.matomo.org/guides/security-in-piwik#checklist)
* [ ] [Wording review done](https://developer.matomo.org/guides/pull-request-reviews#translations-wording-review-done)
* [ ] [Code review done](https://developer.matomo.org/guides/pull-request-reviews#code-review-done)
* [ ] [Tests were added if useful/possible](https://developer.matomo.org/guides/pull-request-reviews#tests-were-added-if-usefulpossible)
* [ ] [Reviewed for breaking changes](https://developer.matomo.org/guides/pull-request-reviews#reviewed-for-breaking-changes)
* [ ] [Developer changelog updated if needed](https://developer.matomo.org/guides/pull-request-reviews#developer-changelog-updated-if-needed)
* [ ] [Documentation added if needed](https://developer.matomo.org/guides/pull-request-reviews#documentation-added-if-needed)
* [ ] Existing documentation updated if needed