@sgiehl 2022-12-19
@tsteur Generally those changes are looking fine. But I was wondering if adding / changing indexes isn't a database operation we wanted to avoid for Matomo 5. We postponed issues like https://github.com/matomo-org/matomo/pull/17466 to Matomo 6, as they would add new dimensions. Just wondering how intense changing indexes is on bigger databases...